requirements.txt
apache-airflow
dbt-snowflake==1.8.0
snowflake-connector-python